bin/d-fine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "d-fine-exporter"
version = "0.1.0"
description = "D-FINE ONNX model exporter with proper dependency management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.1",
    "torchvision>=0.15.2",
    "faster-coco-eval>=1.6.6",
    "PyYAML",
    "tensorboard>=2.0.0",
    "scipy",
    "calflops",
    "transformers",
    "loguru",
    "onnx>=1.12.0",
    "onnxsim>=0.4.0",
    "onnxruntime>=1.15.0",
    "pillow",
    "numpy",
    "opencv-python",
]

[tool.hatch.build.targets.wheel]
packages = []
